requirements.txt
claude-agent-sdk>=0.1.39,<0.2.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
websockets>=13.0
python-multipart>=0.0.17
psutil>=6.0.0
aiofiles>=24.0.0
apscheduler>=3.10.0,<4.0.0
pywinpty>=2.0.0; sys_platform == "win32"
pyyaml>=6.0.0
python-docx>=1.1.0
openpyxl>=3.1.0
PyPDF2>=3.0.0
python-pptx>=1.0.0

# Dev dependencies
ruff>=0.8.0
mypy>=1.13.0
pytest>=8.0.0
types-PyYAML>=6.0.0
